Refactor actions and reducers
