+++ app/frontend/javascripts/categories/categories.reducers.js
@@ -1,0 +1,9 @@
+import { FETCH_CATEGORIES } from './categories.actions';
+
+export default function (state = [], action) {
+  switch (action.type) {
+    case FETCH_CATEGORIES:
+      return action.payload.categories
+  }
+  return state;
+}
