+++ app/frontend/javascripts/districts/districts.reducers.js
@@ -1,0 +1,9 @@
+import { FETCH_DISTRICTS } from './districts.actions';
+
+export default function (state = [], action) {
+  switch (action.type) {
+    case FETCH_DISTRICTS:
+      return action.payload.districts
+  }
+  return state;
+}
